growda/backend/requirements.txt
tensorflow>=2.10.0
fastapi>=0.95.0
uvicorn>=0.21.0
flwr>=1.4.0
pycryptodome>=3.17.0
python-multipart>=0.0.6
numpy>=1.23.5
pillow>=9.4.0